CHORE: move signer locgic out of cmd
