+++ pkg/signer/spiffe/spiffe.go
